+++ contractsSol5/IKyberReserve.sol
@@ -5,18 +5,19 @@ import "./IERC20.sol";
 
 /// @title Kyber Reserve contract
 interface IKyberReserve {
-
     function trade(
         IERC20 srcToken,
-        uint srcAmount,
+        uint256 srcAmount,
         IERC20 destToken,
         address payable destAddress,
-        uint conversionRate,
+        uint256 conversionRate,
         bool validate
-    )
-        external
-        payable
-        returns(bool);
+    ) external payable returns (bool);
 
-    function getConversionRate(IERC20 src, IERC20 dest, uint srcQty, uint blockNumber) external view returns(uint);
+    function getConversionRate(
+        IERC20 src,
+        IERC20 dest,
+        uint256 srcQty,
+        uint256 blockNumber
+    ) external view returns (uint256);
 }
