+++ contractsSol5/ISimpleKyberProxy.sol
@@ -6,7 +6,21 @@ import "./IERC20.sol";
 /// @title simple Kyber Network proxy interface
 /// add convenient functions to help with kyber proxy API
 interface ISimpleKyberProxy {
-    function swapTokenToToken(IERC20 src, uint srcAmount, IERC20 dest, uint minConversionRate) external returns(uint destAmount);
-    function swapEtherToToken(IERC20 token, uint minConversionRate) external payable returns(uint destAmount);
-    function swapTokenToEther(IERC20 token, uint srcAmount, uint minConversionRate) external returns(uint destAmount);
+    function swapTokenToToken(
+        IERC20 src,
+        uint256 srcAmount,
+        IERC20 dest,
+        uint256 minConversionRate
+    ) external returns (uint256 destAmount);
+
+    function swapEtherToToken(IERC20 token, uint256 minConversionRate)
+        external
+        payable
+        returns (uint256 destAmount);
+
+    function swapTokenToEther(
+        IERC20 token,
+        uint256 srcAmount,
+        uint256 minConversionRate
+    ) external returns (uint256 destAmount);
 }
